Rename package pipeline -> types
