--- pipeline/decoder.go
@@ -1,7 +1,0 @@
-package pipeline
-
-// Decoder is the component that parses a raw buffer to a structure
-type Decoder interface {
-	Init(int) error
-	Decode(Messenger) error
-}
