--- pipeline/encoder.go
@@ -1,7 +1,0 @@
-package pipeline
-
-// Encoder serializes a data structure to a output buffer
-type Encoder interface {
-	Init(int) error
-	Encode(Messenger) error
-}
