+++ types/messenger.go
@@ -1,4 +1,4 @@
-package pipeline
+package types
 
 // Messenger is used by modules to handle messages
 type Messenger interface {
